refactor(dash): move socket to socket layout
